+++ tests/deployments/test_cli.py
@@ -151,50 +151,3 @@ def test_explain_with_no_target_implementation(tmp_path):
         )
         assert type(res) == MlflowException
         mock_explain.assert_called_once()
-
-
-def test_start_help():
-    runner = CliRunner()
-    res = runner.invoke(
-        cli.start_server,
-        ["--help"],
-        catch_exceptions=False,
-    )
-    assert res.exit_code == 0
-
-
-def test_start_invalid_config(tmp_path):
-    runner = CliRunner()
-    config = tmp_path.joinpath("config.yml")
-    res = runner.invoke(
-        cli.start_server,
-        ["--config-path", config],
-        catch_exceptions=False,
-    )
-    assert res.exit_code == 2
-    assert "does not exist" in res.output
-
-    config.write_text("\t")
-    res = runner.invoke(
-        cli.start_server,
-        ["--config-path", config],
-        catch_exceptions=False,
-    )
-    assert res.exit_code == 2
-    assert "not a valid yaml file" in res.output
-
-    config.write_text(
-        """
-routes:
-  - model:
-    name: invalid
-"""
-    )
-    res = runner.invoke(
-        cli.start_server,
-        ["--config-path", config],
-        catch_exceptions=False,
-    )
-    assert res.exit_code == 2
-    assert "The gateway configuration is invalid" in res.output
-    assert "routes" in res.output
